pyproject.toml
[tool.poetry]
name = "app"
version = "0.1"
description = "osapiens Makeathon API"
authors = ["Viacheslav Sydora"]

[tool.poetry.dependencies]
python = "^3.10.4"
python-multipart = "^0.0.9"
fastapi = "^0.97.0"
FastAPI-SQLAlchemy = "^0.2.1"
orjson = "^3.9.1"
pydantic = "^1.10.9"
python-dotenv = "^0.21.1"
requests = "^2.31.0"
SQLAlchemy = "^2.0.16"
uvicorn = "^0.18.3"
black = "^23.3.0"
alembic = "^1.11.1"
pytest = "^7.3.2"
unittest2 = "^1.1.0"
loguru = "^0.7.0"
tenacity = "^8.2.2"
psycopg = "^3.1.9"
pre-commit = "^3.3.3"
pillow = "^10.3.0"
numpy = "^1.26.0"
python-jose = "^3.3.0"
